Make storm track optional

+++ 2_process/src/extract_storm_line.R
@@ -10,16 +10,20 @@ extract_storm_line <- function(ind_file, storm_shp_ind, cfg){
   # get the shp zip file from Drive if we don't already have it
   storm_shp_file <- sc_retrieve(storm_shp_ind)
 
-  # unzip into a temporary directory and read from there
-  tmp_path <- file.path(tempdir(), 'storm_track')
-  if (!dir.exists(tmp_path)) dir.create(tmp_path)
-  unzip(storm_shp_file, exdir = tmp_path)
-  layer <- tools::file_path_sans_ext(list.files(tmp_path, pattern='lin\\.shp$'))
-  track <- sf::read_sf(tmp_path, layer=layer)
-  unlink(tmp_path, recursive = TRUE)
+  if(file.size(storm_shp_file) == 0) {
+    track_proj <- NULL
+  } else {
+    # unzip into a temporary directory and read from there
+    tmp_path <- file.path(tempdir(), 'storm_track')
+    if (!dir.exists(tmp_path)) dir.create(tmp_path)
+    unzip(storm_shp_file, exdir = tmp_path)
+    layer <- tools::file_path_sans_ext(list.files(tmp_path, pattern='lin\\.shp$'))
+    track <- sf::read_sf(tmp_path, layer=layer)
+    unlink(tmp_path, recursive = TRUE)
 
-  # convert to the vizzy projection
-  track_proj <- sf::st_transform(track, crs=sf::st_crs(cfg$projection))
+    # convert to the vizzy projection
+    track_proj <- sf::st_transform(track, crs=sf::st_crs(cfg$projection))
+  }
 
   # write the data and return an indicator file
   data_file <- as_data_file(ind_file)
